\name{shoulderToShoulder}
\alias{shoulderToShoulder}

\title{shoulderToShoulder}

\description{A function that stacks a matrix next to itself a specified number of times.
}

\usage{shoulderToShoulder(x)}

\arguments{
  \item{x}{
  n: specifies how many times to stack the matrix.
  bMat: specifies the matrix.
}
}

\author{Katie Monroe}
